Migration to Platform.IO, split of README and CHANGELOG
