+++ .gitignore
@@ -30,3 +30,10 @@
 *.exe
 *.out
 *.app
+
+# PlatformIO & VSCode IDE temp files
+.pio
+.vscode/.browse.c_cpp.db*
+.vscode/c_cpp_properties.json
+.vscode/launch.json
+.vscode/ipch
