+++ .vscode/spellright.dict
@@ -1,0 +1,31 @@
+Nightscout
+INI
+DNS
+IP
+Dexcom
+Sugarmate
+CHANGELOG
+md
+glycemia
+Libre
+Miao
+backlight
+Glimp
+DHT
+ENV
+BTC
+config
+mins
+grey
+Arduino
+JSON
+IDE
+Aliexpress
+USD
+SSID
+jpg
+MERCHANTABILITY
+LGPL
+GPL
+SGV
+startup
