--- M5NSWebConfig.h
@@ -1,12 +1,0 @@
-#ifndef _M5NSWEBCONFIG_H
-#define _M5NSWEBCONFIG_H
-
-void handleRoot();
-void handleUpdate();
-void handleSwitchConfig();
-void handleEditConfigItem();
-void handleGetEditConfigItem();
-void handleSaveConfig();
-void handleNotFound();
-
-#endif
