+++ platformio.ini
@@ -1,0 +1,27 @@
+;PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+[platformio]
+description = M5 Nightscout Monitor
+default_envs = m5stack-core-esp32
+
+[env:m5stack-core-esp32]
+platform = espressif32
+board = m5stack-core-esp32
+framework = arduino
+
+lib_deps =
+  M5Stack
+  ArduinoJson
+
+lib_ignore =
+  Time ; to prevent Time_ID44 lib conflicting with stdc time.h 
+
+build_flags =
+  -Dgpio_deep_sleep_hold_dis='_NOP' ; fix for m5stack core missing this method
